experiments/langchain-kg-agent/requirements.txt
neo4j
langchain
langchain-ollama>=0.1
langchain-anthropic
langchain-core
python-dotenv
